removed memory db and moved types to models
